moved DoubleSpeedcontroller to utils
